+++ middlewares/uploadImage.js
@@ -1,16 +1,9 @@
 const multer = require('multer')
 const fs = require('fs')
-const util = require('util')
-const { PutObjectCommand } = require('@aws-sdk/client-s3')
+
 const path = require('path')
 const { catchAsync } = require('../helpers/catchAsync')
 const ApiError = require('../helpers/ApiError')
-
-const unlinkFile = util.promisify(fs.unlink)
-const { s3Client } = require('../libs/s3Client')
-
-const BUCKET = process.env.S3_BUCKET_NAME
-const REGION = process.env.AWS_REGION
 
 const upload = multer({
   fileFilter: (req, file, cb) => {
@@ -26,29 +19,30 @@ const upload = multer({
     },
     filename: (req, file, cb) => {
       const name = req.body.name || req.body.email
-      cb(null, `${name}-${Date.now()}${path.extname(file.originalname)}`)
+      cb(null, `${name || 'slide'}-${Date.now()}${path.extname(file.originalname)}`)
     },
   }),
 })
 
 module.exports = {
+  /**
+   * Middleware to upload image to temporary folder
+   *
+   * @param {string} image the name of the key in the req.body
+   * @return {void} if no error occurs the image is saved in the temporary folder
+   * and the path is saved in req.body.${image} else an error is thrown
+   */
   uploadImage: (image) => catchAsync(async (req, res, next) => {
     upload.single(image)(req, res, async (err) => {
       if (err) {
         return next(new ApiError(400, err.message))
       }
-      const fileStream = fs.createReadStream(req.file.path)
-      const uploadParams = {
-        Bucket: BUCKET,
-        Body: fileStream,
-        Key: req.file.filename,
+      if (!req.file) {
+        return next()
       }
 
-      await s3Client.send(new PutObjectCommand(uploadParams))
-
-      const fileLocation = `https://${BUCKET}.s3.${REGION}.amazonaws.com/${req.file.filename}`
-      req.file.location = fileLocation
-      await unlinkFile(req.file.path)
+      // eslint-disable-next-line no-eval
+      eval(`req.body.${image} = '${req.file.path}'`)
 
       return next()
     })
